src/EMRNext.Core/Infrastructure/Reporting/ReportingEngine.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.Extensions.Logging;

namespace EMRNext.Core.Infrastructure.Reporting
{
    public class ReportingEngine
    {
        // Core Reporting Interfaces
        public interface IReportDataSource
        {
            Task<IQueryable<object>> GetDataAsync();
        }

        public interface IReportFilter
        {
            IQueryable<object> ApplyFilter(IQueryable<object> data);
        }

        public interface IReportAggregator
        {
            object Aggregate(IQueryable<object> data);
        }

        // Comprehensive Reporting Configuration
        public class ReportConfiguration
        {
            public string ReportId { get; set; }
            public string ReportName { get; set; }
            public ReportType Type { get; set; }
            public List<ReportParameter> Parameters { get; set; }
            public List<string> RequiredPermissions { get; set; }
        }

        // Report Parameter Model
        public class ReportParameter
        {
            public string Name { get; set; }
            public Type ParameterType { get; set; }
            public bool IsRequired { get; set; }
            public object DefaultValue { get; set; }
        }

        // Report Generation Context
        public class ReportGenerationContext
        {
            public ReportConfiguration Configuration { get; set; }
            public Dictionary<string, object> InputParameters { get; set; }
            public string GeneratedBy { get; set; }
            public DateTime GeneratedAt { get; set; }
        }

        // Report Types Enumeration
        public enum ReportType
        {
            Standard,
            Summary,
            Detailed,
            Comparative,
            Trend,
            Predictive
        }

        // Report Result Model
        public class ReportResult
        {
            public string ReportId { get; set; }
            public ReportType Type { get; set; }
            public object RawData { get; set; }
            public object ProcessedData { get; set; }
            public Dictionary<string, object> Metadata { get; set; }
            public List<ReportValidationIssue> ValidationIssues { get; set; }
        }

        // Validation Issue Model
        public class ReportValidationIssue
        {
            public string Code { get; set; }
            public string Message { get; set; }
            public ReportValidationSeverity Severity { get; set; }
        }

        // Validation Severity Enum
        public enum ReportValidationSeverity
        {
            Information,
            Warning,
            Error
        }

        // Core Reporting Service
        public class ReportingService
        {
            private readonly ILogger<ReportingService> _logger;
            private readonly List<IReportDataSource> _dataSources;
            private readonly List<IReportFilter> _filters;
            private readonly List<IReportAggregator> _aggregators;

            public ReportingService(
                ILogger<ReportingService> logger,
                IEnumerable<IReportDataSource> dataSources,
                IEnumerable<IReportFilter> filters,
                IEnumerable<IReportAggregator> aggregators)
            {
                _logger = logger;
                _dataSources = dataSources.ToList();
                _filters = filters.ToList();
                _aggregators = aggregators.ToList();
            }

            // Generate Report
            public async Task<ReportResult> GenerateReportAsync(
                ReportGenerationContext context)
            {
                try
                {
                    // Validate input parameters
                    ValidateReportParameters(context);

                    // Fetch data from sources
                    var data = await FetchDataAsync(context);

                    // Apply filters
                    data = ApplyFilters(data, context);

                    // Aggregate data
                    var aggregatedData = AggregateData(data, context);

                    // Create report result
                    return new ReportResult
                    {
                        ReportId = context.Configuration.ReportId,
                        Type = context.Configuration.Type,
                        RawData = data,
                        ProcessedData = aggregatedData,
                        Metadata = CreateReportMetadata(context, data),
                        ValidationIssues = new List<ReportValidationIssue>()
                    };
                }
                catch (Exception ex)
                {
                    _logger.LogError(ex, "Report generation failed");
                    return CreateErrorReport(ex);
                }
            }

            // Validate Report Parameters
            private void ValidateReportParameters(ReportGenerationContext context)
            {
                foreach (var param in context.Configuration.Parameters)
                {
                    if (param.IsRequired && 
                        !context.InputParameters.ContainsKey(param.Name))
                    {
                        throw new ArgumentException(
                            $"Required parameter {param.Name} is missing"
                        );
                    }
                }
            }

            // Fetch Data from Sources
            private async Task<IQueryable<object>> FetchDataAsync(
                ReportGenerationContext context)
            {
                var combinedData = new List<object>();

                foreach (var dataSource in _dataSources)
                {
                    var sourceData = await dataSource.GetDataAsync();
                    combinedData.AddRange(sourceData);
                }

                return combinedData.AsQueryable();
            }

            // Apply Filters
            private IQueryable<object> ApplyFilters(
                IQueryable<object> data, 
                ReportGenerationContext context)
            {
                foreach (var filter in _filters)
                {
                    data = filter.ApplyFilter(data);
                }
                return data;
            }

            // Aggregate Data
            private object AggregateData(
                IQueryable<object> data, 
                ReportGenerationContext context)
            {
                foreach (var aggregator in _aggregators)
                {
                    return aggregator.Aggregate(data);
                }
                return data;
            }

            // Create Report Metadata
            private Dictionary<string, object> CreateReportMetadata(
                ReportGenerationContext context, 
                IQueryable<object> data)
            {
                return new Dictionary<string, object>
                {
                    { "TotalRecords", data.Count() },
                    { "GeneratedAt", DateTime.UtcNow },
                    { "GeneratedBy", context.GeneratedBy }
                };
            }

            // Create Error Report
            private ReportResult CreateErrorReport(Exception ex)
            {
                return new ReportResult
                {
                    ValidationIssues = new List<ReportValidationIssue>
                    {
                        new ReportValidationIssue
                        {
                            Code = "REPORT_GENERATION_ERROR",
                            Message = ex.Message,
                            Severity = ReportValidationSeverity.Error
                        }
                    }
                };
            }
        }
    }
}
